ajuste no valor da saida do exit
